+++ src/main/java/com/zju/vis/print_backend/vo/ExcelRelDateRawMaterialVo.java
@@ -1,8 +1,6 @@
 package com.zju.vis.print_backend.vo;
 
-import com.alibaba.excel.annotation.ExcelProperty;
-import com.alibaba.excel.annotation.format.DateTimeFormat;
-import com.zju.vis.print_backend.convert.DateConvert;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -20,5 +18,5 @@ public class ExcelRelDateRawMaterialVo {
     // @DateTimeFormat("yyyy.MM.dd")
     private String rawMaterialDate;
     // 当期价格
-    private Float price;
+    private Double price;
 }
